refactor: allow all clients to make broker admin requests
This moves the type from an internal type in the gRPC gateway
to a shared type in the broker client to allow reuse without
depending on the entire gateway.
