+++ zeebe/broker-client/src/main/java/io/camunda/zeebe/broker/client/api/dto/BrokerAdminRequest.java
@@ -5,11 +5,9 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.admin;
+package io.camunda.zeebe.broker.client.api.dto;
 
 import io.atomix.cluster.BrokerMemberId;
-import io.camunda.zeebe.broker.client.api.dto.BrokerRequest;
-import io.camunda.zeebe.broker.client.api.dto.BrokerResponse;
 import io.camunda.zeebe.protocol.Protocol;
 import io.camunda.zeebe.protocol.impl.encoding.AdminRequest;
 import io.camunda.zeebe.protocol.impl.encoding.AdminResponse;
